+++ HomeBarSwiftUI/HomeBarSwiftUI/GoogleService-Info.plist
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>731945684121-kb2e8kl548njl4mcuqjqu3si22rl7opr.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.731945684121-kb2e8kl548njl4mcuqjqu3si22rl7opr</string>
+	<key>API_KEY</key>
+	<string>AIzaSyAla0nfUJi3cejQG5lWfno9wX3RC9FLGiY</string>
+	<key>GCM_SENDER_ID</key>
+	<string>731945684121</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>nu-shtosh.HomeBarSwiftUI</string>
+	<key>PROJECT_ID</key>
+	<string>homebar-dream</string>
+	<key>STORAGE_BUCKET</key>
+	<string>homebar-dream.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:731945684121:ios:467d189c221d7003e0bc08</string>
+</dict>
+</plist>
