+++ HomeBarSwiftUI/HomeBarSwiftUI/Persistence.swift
@@ -22,8 +22,8 @@ struct PersistenceController {
         } catch {
             // Replace this implementation with code to handle the error appropriately.
             // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
-            let nsError = error as NSError
-            fatalError("Unresolved error \(nsError), \(nsError.userInfo)")
+            let NSError = error as NSError
+            fatalError("Unresolved error \(NSError), \(NSError.userInfo)")
         }
         return result
     }()
